+++ modulo_tag.go
@@ -1,19 +1,39 @@
 package main
 
 import (
+	"encoding/json"
+	"fmt"
 	"log"
 	"strconv"
-
 	"github.com/gin-gonic/gin"
 )
 
 // --------------- Etiquetas ----------------------------------------
 
-// TO DO: cambiar la consulta (AQUÍ SE SACAN TODAS LAS ETIQUETAS DE UN USUARIO)
 func GetTagsByUserId(c *gin.Context) {
 
 	//ID del usuario
 	id := c.Param("id")
+
+	//	Consulta a redis
+	val, err := rdb.Get(c.Request.Context(), "TagsByUser:"+id).Result()
+
+	if err == nil {
+		fmt.Printf("\n Si existe registro")
+		var tagsArray []Tags
+
+		err := json.Unmarshal([]byte(val), &tagsArray)
+
+		if err == nil {
+			c.JSON(200, tagsArray)
+			return
+
+		}
+
+	}
+
+	// Si no existe en redis, se debe crear la consulta
+	fmt.Printf("\n>>>>Creando registro")
 
 	rows, err := db.Query(`
 		SELECT * FROM EtiquetasRecordatorios 
@@ -55,10 +75,11 @@ func GetTagsByUserId(c *gin.Context) {
 		return
 	}
 
+	// Devuelve la consulta de la base relacional
 	c.JSON(200, TagsArray)
 }
 
-// TO DO: FUNCION PARA SACAR LAS ETIQUETAS DE UN RECORDATORIO POR SU NOMBRE
+// FUNCION PARA SACAR LAS ETIQUETAS DE UN RECORDATORIO POR SU NOMBRE
 func GetTagsByUserIdAndReminderId(c *gin.Context) {
 
 	//ID del usuario
@@ -71,6 +92,27 @@ func GetTagsByUserIdAndReminderId(c *gin.Context) {
 		c.JSON(400, gin.H{"error": "Invalid reminder id"})
 		return
 	}
+
+	//	Consulta a redis
+	key := fmt.Sprintf("TagsByUser&Reminder:%s-%d", id, reminderId)
+	val, err := rdb.Get(c.Request.Context(), key).Result()
+
+	if err == nil {
+		fmt.Printf("\n Si existe registro")
+		var tagsArray []Tags
+
+		err := json.Unmarshal([]byte(val), &tagsArray)
+
+		if err == nil {
+			c.JSON(200, tagsArray)
+			return
+
+		}
+
+	}
+
+	// Si no existe en redis, se debe crear la consulta
+	fmt.Printf("\n>>>>Creando registro")
 
 	rows, err := db.Query(`
 		SELECT * FROM EtiquetasRecordatorios 
@@ -112,10 +154,11 @@ func GetTagsByUserIdAndReminderId(c *gin.Context) {
 		return
 	}
 
+	// Devuelve la consulta de la base relacional
 	c.JSON(200, TagsArray)
 }
 
-// TO DO: DELETE TAG
+// DELETE TAG
 func deleteTag(c *gin.Context) {
 
 	var delTag DelTag
@@ -125,7 +168,24 @@ func deleteTag(c *gin.Context) {
 		c.JSON(400, gin.H{"error": "formato invalido de json"})
 		return
 	}
+	if !AuthorityCheck(*delTag.CodUsuario, c) {
+		c.AbortWithStatusJSON(401, gin.H{"error": "Autorización requerida"})
+		return
+	}
 
+	// Borrar registro de etiquetas de usuario de redis
+	deleted, err2 := rdb.Del(ctx, "TagsByUser:"+*delTag.CodUsuario).Result()
+
+	if err2 != nil {
+		fmt.Printf("\nError de conexión: %v", err2)
+
+	} else if deleted > 0 {
+		fmt.Printf("\nRegistro eliminado con éxito")
+	} else {
+		fmt.Printf("\nNo se encontró registro relacionado")
+	}
+
+	// Llamado al procedimiento
 	result, err := db.Exec("CALL eliminar_etiqueta(?)", delTag.N_idEtiqueta)
 
 	if err != nil {
@@ -133,6 +193,12 @@ func deleteTag(c *gin.Context) {
 		c.JSON(500, gin.H{"error": "Internal server error"})
 		return
 	}
+
+	descripcion := "Se eliminó/recuperó etiqueta | ID: " +
+		strconv.Itoa(delTag.N_idEtiqueta) +
+		" | Usuario ID: " + strconv.Itoa(delTag.P_usuario)
+
+	insertarLog(delTag.P_usuario, "ELIMINAR_ETIQUETA", descripcion)
 
 	rowsAffected, _ := result.RowsAffected()
 	c.JSON(200, gin.H{
